Play a game of five rounds
re-wrote the player input so that it is run by a function, rather than a variable
Re-wrote both the player input function and the computer input function so they use similar variables to one another.
Wrote variables setting the score to zero.
Re-wrote the playRound variable to take into account the new playerPlay and computerPlay variables, declared in the playGame function.
Re-wrote the playRound function to add to the score with playerScore++, tieScore++ and computerScore++.
Included the playGame function, which declares the playerPlay and computerPlay variables, then runs the playRound function five times.